Model two-state CAS as LL/SC internally
This'll make emitting assembly on certain architectures easier.

To retain the ability to emit a single CAS instruction, I currently
stuff the bits of the SC that are relevant for CAS building into the LL.
This is slightly semantically weird, but the other possibilities for
carrying that information to the C emitter seem equally unappealing.