Mostly fix overly eager semicolon insertion
We now check each basic instruction to see if it needs a semicolon.
StoreCnd and the transaction instructions don't, as they either emit
bits of block syntax or a comment.

(NB: the comment emitted on StoreCnd is a bit of a kludge; it's there to
prevent either a blank line or a double indent from showing up.  Feel
free to remove/change/improve!)